send data to cboard

+++ io/command.hpp
@@ -3,13 +3,13 @@
 
 namespace io
 {
-  struct Command
-  {
-    int state;
-    double x_bias;
-    double y_bias;
-  };
+struct Command
+{
+  int state;
+  int x_bias;
+  int y_bias;
+};
 
-} // namespace io
+}  // namespace io
 
-#endif // IO__COMMAND_HPP
+#endif  // IO__COMMAND_HPP
